Rename auto-sleep.sh to sleep_when_lid_closed.sh
Fixed an issue in start_if_absent.sh.  Previously we looked for
process name (equivalent to "-o ucmd") which has the advantage that we
won't hit the search string passed as the argument of the script and
grep.  However, the process name truncated by kernel limit (approx. 16
chars).  Now we use "-o command" instead to get the full command line,
although we have to filter out grep and the start_if_absent.sh script.